requirements.txt
Django==5.0.6
djangorestframework==3.15.2
drf-spectacular==0.27.2
psycopg2-binary==2.9.9
python-decouple==3.8
structlog==24.2.0
gunicorn==22.0.0
pytest-django==4.8.0
factory-boy==3.3.0
pytest==8.2.2
pytest-cov==5.0.0
